pizza-chatbot/requirements.txt
streamlit
langchain
langchain-ollama
langchain-core
qrcode[pil]
pillow
